% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{bootstrap test}
\usage{
bootstrap(x, theta0, b)
}
\arguments{
\item{x}{data}

\item{theta0}{null value of the mean}

\item{b}{the number of bootstrap resamples}
}
\value{
origtest pvalue teststatall
}
\description{
test the mean  using bootstrap
}
\examples{
bootstrap(rnorm(200,mean=0.9,sd=6),0.6,50)
}
