% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixnormal.R
\name{mixnormal}
\alias{mixnormal}
\title{EM algorithm}
\usage{
mixnormal(x, theta0)
}
\arguments{
\item{x}{data}

\item{theta0}{unknown parameter}
}
\value{
estimated value
}
\description{
Estimating the parameters of a mixed normal model using R
}
\examples{
x<- rnorm(100,mean=2,sd=2)
theta0<- c(1,1,2,3,0.5)
mixnormal(x,theta0)
}
