% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piest.R
\name{piest}
\alias{piest}
\title{Estimate the value of pi}
\usage{
piest(n)
}
\arguments{
\item{n}{the number of simulations}
}
\value{
value of estimate and it's error
}
\description{
The package is used for estimating pi.
}
\examples{
piest(10^6)
}
